requirements.txt
# reg-change-engine dependencies

# HTTP & env
requests>=2.31.0
python-dotenv>=1.0.0

# XML parsing (stdlib ET + lxml for robustness)
lxml>=5.1.0

# Diff / NLP
transformers>=4.40.0
torch>=2.2.0

# Database
psycopg2-binary>=2.9.9

# API server
fastapi>=0.111.0
uvicorn[standard]>=0.29.0

# Pydantic (FastAPI peer dep)
pydantic>=2.7.0
